Finish the owl
